+++ CustomModules/Nigrimmist.Modules/Modules/ToolUpdate.cs
@@ -73,7 +73,7 @@ namespace Nigrimmist.Modules.Modules
                 AppVersionInfo info = JsonConvert.DeserializeObject<AppVersionInfo>(versionJson);
                 var versions = info.Versions.OrderBy(x => x.Version);
                 var latestVersion = versions.Last();
-                if (latestVersion.Version > (float)_client.UiClientVersion)
+                if (latestVersion.Version > _client.UiClientVersion)
                 {
                     string message = "Версия v" + latestVersion.Version + " доступна для скачивания. " + Environment.NewLine + "Кликните сюда для перехода на новую версию" + Environment.NewLine+Environment.NewLine+"В новой версии :" + Environment.NewLine + Environment.NewLine + latestVersion.WhatsNew;
                     _client.ShowMessage(eventToken, message, "Вышла новая версия.").OnClick(() =>
